CD4(gamlss.data)	R Documentation

The CD4 Count Data files for GAMLSS

Description

CD4: The data were given by Wade and Ades (1994) and refer to cd4 counts from uninfected children born to HIV-1 mothers and the age of the child.

Usage



data(CD4)

Format

Data frames each with the following variable.

cd4

a numeric vector showing the CD4 counts

age

the age of the child

Details

Data sets usefull for the GAMLSS booklet

References

Wade, A. M. and Ader, A. E. (1994) Age-related reference ranges : Significance tests for models and confidence intervals for centiles. Statistics in Medicine, 13, pages 2359-2367.

Examples


data(CD4)
with(CD4,plot(cd4~age))

[Package gamlss.data version 0.0-2 Index]