LGAclaims(gamlss.data)	R Documentation

The LGA Claims Data files for GAMLSS

Description

These are several small data files usefull for gamlss fits.

LGAclaims: the data were given by Gillian Heller and can be found in de Jong and Heller (2007). This data set records the number of third party claims, Claims, in a twelve month period between 1984-1986 in each of 176 geographical areas (local government areas) in New South Wales, Australia. Areas are grouped into thirteen statistical divisions (SD). Other recorded variables are the number of accidents, Accidents, the number of people killed or injured and population with all variables classified according to area.

Usage




data(LGAclaims)

Format

Data frames each with the following variable.

Claims

the number of third party claims

LGA

Local government areas in New South Wales

SD

statistical divisions

Pop$_$density

population density

KI

the number of people killed or injured

Accidents

the number of accidents

Population

population size

L$_$KI

log of KI

L$_$Accidents

the log of the number of accidents

L$_$Population

log Population

Details

Data sets usefull for the GAMLSS booklet

References

de Jong, P. and Heller G. (2007) Generalized Linear Models for Insurance Data , Cambridge University Press

Examples


data(LGAclaims)
with(LGAclaims, plot(data.frame(Claims, Pop_density, KI, Accidents, Population)))

[Package gamlss.data version 0.0-2 Index]