MX.control(gamlss.mx)	R Documentation

The control function for gamlssMX

Description

The function sets controls for the gamlssMX function.

Usage


MX.control(cc = 1e-04, n.cyc = 200, trace = FALSE, 
        seed = NULL, plot = TRUE, sample = NULL, ...)

Arguments

cc 	 convergent criterion for the EM

n.cyc 	 number of cycles for EM

trace 	 whether to print the EM iterations

seed 	 a number for setting the seeds for starting values

plot 	 whether to plot the sequence of global deviance up to convergence

sample 	 how large the sample to be in the starting values

... 	 for extra arguments

Value

Returns a list

Author(s)

Mikis Stasinopoulos and Bob Rigby

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape, (with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

See Also

gamlss, gamlssMX, gamlssMXfits

[Package gamlss.mx version 2.0-0 Index]