Mums(gamlss.util)	R Documentation

Mothers encouragement data

Description

Mothers encouragement for participation in Higher Education. The response variable is mums a three level factor which can be used in a multinomial Logistic model or mumsB a two level factor suitable for binary logistic model.

Usage

data(Mums)

Format

A data frame with 871 observations on the following 7 variables.

mums

mothers encouragement: factor with levels 1 is for strong encouragement, 2 is for some encouragement and 3 for no encouragement/discouragement

class

social class: a factor with levels 1is C1, 2 is C2, 3 is D and 4 is E

age

age of the participants: a factor with levels 1 is 16-18, 2 is 19-20 and 3 is 20-30

gender

a factor with levels 1 is male and 2 is female

ethn

ethnicity of the participants: a factor with levels 1 is white, 2 is black, 3 is asian and 4 is other

qual

qualifications of the participants: a factor with levels, 1 is greater or equal to 2 A levels, 2 is HND or more than 5 GCSE's, 3 is less than 5 GSCSE's ar none above and 4 no formal qualification

mumsb

mothers encouragement: a factor with levels, 0 is no encouragement or some encouragement 1 is for strong encouragement

Details

The data were collected as part of the Social Class and widening Participation in Higher Education Project based at the University of North London (now London Metropolitan University) and supported by the University's Development and Diversity Fund over the period 1998-2000.

Source

Professor Robert Gilchrist director of STORM at London Metropolitan

References

Collier T., Gilchrist R. and Phillips D. (2003), Who Plans to Go to University? Statistical Modelling of potential Working-Class Participants, Education Research and Evaluation, Vol 9, No 3, pp 239-263.

Examples


data(Mums)
MM<-xtabs(~mums+qual, data=Mums)
mosaicplot(MM, color=TRUE)
MM<-xtabs(~mums+ethn+gender, data=Mums)
mosaicplot(MM, color=TRUE)

[Package gamlss.util version 0.0-2 Index]