NL.control(gamlss.nl)	R Documentation

Auxiliary for Controlling non linear GAMLSS Fitting

Description

This ia an auxiliary function used to control the iterations for nlgamlss fitting. Typically only used when calling nlgamlss function with the option control. Since the nlgamlss uses nlm for fitting all of the NL.control argument are passed to nlm.

Usage


NL.control(fscale = 1,  typsize = NULL, stepmax = NULL, iterlim = 100, 
           ndigit = 10, steptol = 1e-05, 
           gradtol = 1e-05, print.level = 0, check.analyticals = TRUE, 
           hessian = TRUE)

Arguments

fscale 	 an estimate of the size of log-likelihood at the minimum with default equal 1.

typsize 	 this argument is passed to nlm and it is an estimate of the size of each parameter at the minimum. If its value is NULL (the default value) the typsizeis set within the nlgamlss function to typsize=abs(p0) where p0 is the vector containing the starting values of all the parameters to be maximized. p0 is defined within nlgamlss

stepmax 	 this argument is passed to nlm and it is a positive scalar which gives the maximum allowable scaled step length. stepmax is used to prevent steps which would cause the optimization function to overflow, to prevent the algorithm from leaving the area of interest in parameter space, or to detect divergence in the algorithm. stepmax would be chosen small enough to prevent the first two of these occurrences, but should be larger than any anticipated reasonable step. If its value is NULL (the default value) it is defined within nlgamlss as stepmax=sqrt(p0 %*% p0)

iterlim 	 a positive integer specifying the maximum number of iterations to be performed before the program is terminated. The default is 100

ndigit 	 the number of significant digits in the log-likelihood function. The default is 10

steptol 	 A positive scalar providing the minimum allowable relative step length. The defaults is 1e-05

gradtol 	 a positive scalar giving the tolerance at which the scaled gradient is considered close enough to zero to terminate the algorithm. The scaled gradient is a measure of the relative change in log-likelihood in each direction 'p[i]' divided by the relative change in 'p[i]'. The default is 1e-05

print.level 	 this argument determines the level of printing which is done during the minimization process. The default value of '0' means that no printing occurs, a value of '1' means that initial and final details are printed and a value of 2 means that full tracing information is printed.

check.analyticals 	 a logical scalar specifying whether the analytic gradients and Hessians, if they are supplied, should be checked against numerical derivatives at the initial parameter values. This can help detect incorrectly formulated gradients or Hessians.

hessian 	 if TRUE, the hessian of the log likelihood at the maximum is returned ,the default is hessian=TRUE

Details

See the R function nlm and the fist two refernces below for details of the algotithm.

Value

A list with the arguments as components.

Note

This functions supports the function nlgamlss

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

Dennis, J. E. and Schnabel, R. B. (1983) Numerical Methods for Unconstrained Optimization and Nonlinear Equations. Prentice-Hall, Englewood Cliffs, NJ.

Schnabel, R. B., Koontz, J. E. and Weiss, B. E. (1985) A modular system of algorithms for unconstrained minimization. ACM Trans. Math. Software, 11, 419-440.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

See Also

nlgamlss, nlm

Examples




[Package gamlss.nl version 1.6-0 Index]