Q.stats(gamlss)	R Documentation

A function to calculate the Q-statistics

Description

This function calculates and prints the Q-statistics which are useful to test normality of the residuals within a range of an independent variable, for example age in centile estimation, see Royston and Wright (2000).

Usage


Q.stats(obj, xvar = NULL, xcut.points = NULL, n.inter = 10, zvals = TRUE, 
        save = TRUE, ...)

Arguments

obj 	 a GAMLSS object or any other residual vector

xvar 	 a unique explanatory variable

xcut.points 	 the x-axis cut off points e.g. c(20,30). If xcut.points=NULL then the n.inter argument is activated

n.inter 	 if xcut.points=NULL this argument gives the number of intervals in which the x-variable will be split, with default 4

zvals 	 if TRUE the output matix contains the individual z's rather that Q statistics

save 	 whether to save the Q-statistics or not with default equal to TRUE. In this case the functions produce a matrix giving individual Q (or z) statistics and the final aggregate Q's

... 	 for extra arguments

Details

Value

A matrix containing the individual Q's and the aggregate Q-statistics

Note

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk, with contributions from Elaine Borghie

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Royston P. and Wright E. M. (2000) Goodness of fit statistics for the age-specific reference intervals. Statistics in Medicine, 19, pp 2943-2962.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, centiles.split, wp

Examples


data(abdom)
h<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=BCT, data=abdom) 
Q.stats(h,xvar=abdom$x,n.inter=8)
Q.stats(h,xvar=abdom$x,n.inter=8,zvals=FALSE)
rm(h)

[Package gamlss version 1.7-9 Index]