additive.fit(gamlss)	R Documentation

Implementing Backfitting in GAMLSS

Description

This function is not to be used on its own. It is used for backfitting in the GAMLSS fitting algorithms and it is based on the equivalent function written by Trevor Hastie in the gam() S-plus implementation, (Chambers and Hastie, 1991).

Usage


additive.fit(x, y, w, s, who, smooth.frame, maxit = 30, tol = 0.001, 
             trace = FALSE, se = TRUE, ...)

Arguments

x 	 the linear part of the explanatory variables

y 	 the response variable

w 	 the weights

s 	 the matrix containing the smoothers

who 	 the current smoothers

smooth.frame 	 the data frame used for the smoothers

maxit 	 maximum number of iterations in the backfitting

tol 	 the tolerance level for the backfitting

trace 	 whether to trace the backfitting algorithm

se 	 whether standard errors are required

... 	 for extra arguments

Details

This function should not be used on its own

Value

Returns a list with the linear fit plus the smothers

Note

Author(s)

Mikis Stasinopoulos

References

Chambers, J. M. and Hastie, T. J. (1991). Statistical Models in S, Chapman and Hall, London.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss

[Package gamlss version 1.7-9 Index]