aids(gamlss.util)	R Documentation

Aids Cases in England and Wales

Description

The quarterly reported AIDS cases in the U.K. from January 1983 to March 1994 obtained from the Public Health Laboratory Service, Communicable Disease Surveillance Centre, London.

Usage

data(aids)

Format

A data frame with 45 observations on the following 3 variables.

y

the number of quarterly aids cases in England and Wales: a numeric vector

x

time in months from January 1983, 1:45 : a numeric vector

qrt

the quarterly seasonal effect a factor with 4 levels, [1=Q1 (Jan-March), 2=Q2 (Apr-June), 3=Q3 (July-Sept), 4=Q4 (Oct-Dec)]

Details

The counts y can be modelled using a (smooth) Poisson regression model in time x with the quarterly effects i.e. cs(x,df=7)+qrt. Overdispersion persists, so use a Negative Binomial distribution of type I or II. The data also can be used to find a break point in time, see Rigby and Stasinopoulos (1992).

Source

Public Health Laboratory Service, Communicable Disease Surveillance Centre, London.

References

Stasinopoulos, D.M. and Rigby, R. A. (1992). Detecting break points in generalized linear models. Computational Statistics and Data Analysis, 13, 461–471.

Examples


data(aids)
attach(aids)
plot(x,y,pch=21,bg=c("red","green3","blue","yellow")[unclass(qrt)])
detach(aids)

[Package gamlss.util version 0.0-2 Index]