alveolar(gamlss.data)	R Documentation

The Alveolar Data files for GAMLSS

Description

alveolar : alveolar-bronchiolar adenomas data used by Tamura and Young (1987) and also reproduce in Hand et al. (1994), data set 256. The data are the number of mice out of certain number of mice (the binomial denominator) in 23 independent groups, having alveolar-bronchiolar adenomas.

Usage



data(alveolar)


Format

Data frames each with the following variable.

r

a numeric vector showing the number of mice out of n number of mice (the binomial denominator below) in 23 independent groups, having alveolar-bronchiolar adenomas.

n

a numeric vector showing the total number of mice

Details

Data sets usefull for the GAMLSS booklet

References

Hand et al. (1994) A handbook of small data sets. Chapman and Hall, London.

Examples


data(alveolar)
with(alveolar, hist(r/n))

[Package gamlss.data version 0.0-2 Index]