db(gamlss.util)	R Documentation

Head Circumference of Dutch Boys

Description

The data are comming from the Fourth Dutch Growth Study, Fredriks et al. (2000a, 2000b), which is a cross-sectional study that measures growth and development of the Dutch population between the ages 0 and 21 years. The study measured, among other variables, height, weight, head circumference and age for 7482 males and 7018 females. Here we have the only the head circumference of Dutch boys.

Usage

data(db)

Format

A data frame with 7040 observations on the following 2 variables.

head

head circumference

age

age in years

Source

The data were kindly given by professor Stef. van Buuren.

References

Fredriks, A.M. van Buuren, S. Burgmeijer, R.J.F. Meulmeester, J.F. Beuker, R.J. Brugman, E. Roede, M.J. Verloove-Vanhorick, S.P. and Wit, J. M. (2000a), Continuing positive secular change in The Netherlands, 1955-1997, Pediatric Research, 47, 316–323

Fredriks, A.M. van Buuren, S. Wit, J.M. and Verloove-Vanhorick, S. P. (2000b) Body index measurments in 1996-7 compared with 1980, Archives of Childhood Diseases, 82, 107–112

van Buuren and Fredriks M. (2001) Worm plot: simple diagnostic device for modelling growth reference curves. Statistics in Medicine, 20, 1259–1277

Examples


data(db)
attach(db)
plot(age,head)
detach(db)

[Package gamlss.util version 0.0-2 Index]