cens.d(gamlss.cens)	R Documentation

Censored probability density function of a gamlss.family distribution

Description

Creates a probability density function from a current gamlss.family distribution to be used for fitting a censored or interval response variable.

Usage


cens.d(family = "NO", type = c("right", "left", "interval"), ...)

Arguments

family 	 a gamlss.family object, which is used to define the distribution and the link functions of the various parameters. The distribution families supported by gamlss() can be found in gamlss.family and in the package gamlss.dist.

type 	 whether right, left or in interval censoring is required, (right is the default)

... 	 for extra arguments

Details

This function is used to calculate the likelihood function for censored data. This function is not supposed to be used on its own but it is used in function gen.cens.

Value

Returns a modified d family function. The argument of the original function d function are the same.

Note

For an example see gen.cens()

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk and Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

cens.p, gen.cens

Examples

#see the help for function cens for an example

[Package gamlss.cens version 1.8.1 Index]