centiles(gamlss)	R Documentation

Plots the centile curves for a GAMLSS object

Description

This function plots centiles curves for distributions belonging to the GAMLSS family of distributions. The function also tabulates the sample percentages below each centile curve (for comparison with the model percentages given by the argument cent.) A restriction of the function is that it applies to models with one explanatory variable only.

Usage


centiles(obj, xvar = NULL, cent = c(0.4, 2, 10, 25, 50, 75, 90, 98, 99.6), 
         legend = TRUE, ylab = "y", xlab = "x", main = NULL, 
         main.gsub = "@", xleg = min(xvar), yleg = max(obj$y), 
         xlim = range(xvar), ylim = range(obj$y), save = FALSE, 
         plot = TRUE, pch = "+", col = "blue", 
         col.centiles = 1:length(cent) + 2, lty.centiles = 1, lwd.centiles = 1, ...)

Arguments

obj 	 a fitted gamlss object from fitting a gamlss distribution

xvar 	 the unique explanatory variable

cent 	 a vector with elements the % centile values for which the centile curves have to be evaluated

legend 	 whether a legend is required in the plot or not, the default is legent=TRUE

ylab 	 the y-variable label

xlab 	 the x-variable label

main 	 the main title here as character. If NULL the default title "centile curves using NO" (or the relevant distributions name) is shown

main.gsub 	 if the main.gsub (with default "@") appears in the main title then it is substituted with the default title.

xleg 	 position of the legend in the x-axis

yleg 	 position of the legend in the y-axis

xlim 	 the limits of the x-axis

ylim 	 the limits of the y-axis

save 	 whether to save the sample percentages or not with default equal to FALSE. In this case the sample percentages are printed but are not saved

plot 	 whether to plot the centiles. This option is useful for centile.split

pch 	 the character to be used as the default in plotting points see par

col 	 plotting colour see par

col.centiles 	 Plotting colours for the centile curves

lty.centiles 	 line type for the centile curves

lwd.centiles 	 The line width for the centile curves

... 	 for extra arguments

Details

Centiles are calculated using the fitted values in obj and xvar must correspond exactly to the predictor in obj to plot correctly.

col.centiles, lty.centiles and lwd.centiles may be vector arguments and are recycled to the length cent if necessary.

Value

A centile plot is produced and the sample centiles below each centile curve are printed (or saved)

Warning

This function is appropriate only when one continuous explanatory variable is fitted in the model

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk with contribution from Steve Ellison

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, centiles.split , centiles.com

Examples


data(abdom)
h<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=BCT, data=abdom) 
# default plot
centiles(h,xvar=abdom$x)
# control of colours and lines
centiles(h, xvar=abdom$x,  col.cent=c(2,3,4,5,1,5,4,3,2,1), 
              lwd.cent=c(1,1,1,1,2,1,1,1,1))
#Control line types
centiles(h, xvar=abdom$x,  col.cent=1, cent=c(.5,2.5,50,97.5,99.5), 
              lty.centiles=c(3,2,1,2,3),lwd.cent=c(1,1,2,1,1))
# control of the main title
centiles(h, xvar=abdom$x,  main="Abdominal data \n @")
rm(h)

[Package gamlss version 1.9-9 Index]