centiles.com(gamlss)	R Documentation

Comparing centiles from different GAMLSS models

Description

This function compares centiles curves for more than one GAMLSS objects. It is based on the centiles function. The function also tabulates the sample percentages below each centile curve (for comparison with the model percentages given by the argument cent.) A restriction of the function is that it applies to models with one explanatory variable only.

Usage


centiles.com(obj, ..., xvar = NULL, cent = c(0.4, 10, 50, 90, 99.6), 
             legend = TRUE, ylab = "y", xlab = "x", xleg = min(xvar), 
             yleg = max(obj$y), xlim = range(xvar), ylim = NULL, 
             no.data = FALSE, color = TRUE, plot = TRUE)

Arguments

obj 	 a fitted gamlss object from fitting a gamlss continuous distribution

... 	 optionally more fitted GAMLSS model objects

xvar 	 the unique explanatory variable

cent 	 a vector with elements the % centile values for which the centile curves have to be evaluated

legend 	 whether a legend is required in the plot or not, the default is legent=TRUE

ylab 	 the y-variable label

xlab 	 the x-variable label

xleg 	 position of the legend in the x-axis

yleg 	 position of the legend in the y-axis

xlim 	 the limits of the x-axis

ylim 	 the limits of the y-axis

no.data 	 whether the data should plotted, default no.data=FALSE or not no.data=TRUE

color 	 whether the fitted centiles are shown in colour, color=TRUE (the default) or not color=FALSE

plot 	 whether to plot the centiles

Details

Value

Centile plots are produced for the different fitted models and the sample centiles below each centile curve are printed

Warning

This function is appropriate only when one continuous explanatory variable is fitted in the model

Note

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk and Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

See Also

gamlss, centiles , centiles.split

Examples


data(abdom)
h1<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=BCT, data=abdom)
h2<-gamlss(y~lo(x,span=0.4), sigma.formula=~lo(x,span=0.4), family=BCT, data=abdom ) 
centiles.com(h1,h2,xvar=abdom$x)
rm(h1,h2)

[Package gamlss version 1.6-5 Index]