centiles.split(gamlss)	R Documentation

Plots centile curves split by x for a GAMLSS object

Description

This function plots centiles curves for separate ranges of the unique explanatory variable x. It is similar to the centiles function but the range of x is split at a user defined values xcut.point into r separate ranges. The functions also tabulates the sample percentages below each centile curve for each of the r ranges of x (for comparison with the model percentage given by cent) The model should have only one explanatory variable.

Usage


centiles.split(obj, xvar = NULL, xcut.points = NULL, n.inter = 4, 
               cent = c(0.4, 2, 10, 25, 50, 75, 90, 98, 99.6), 
               legend = FALSE, main = NULL, main.gsub = "@", 
               ylab = "y", xlab = "x", ylim = NULL, overlap = 0, 
               save = TRUE, plot = TRUE, ...)

Arguments

obj 	 a fitted gamlss object from fitting a gamlss continuous distribution

xvar 	 the unique explanatory variable

xcut.points 	 the x-axis cut off points e.g. c(20,30). If xcut.points=NULL then the n.inter argument is activated

n.inter 	 if xcut.points=NULL this argument gives the number of intervals in which the x-variable will be splited, with default 4

cent 	 a vector with elements the % centile values for which the centile curves are to be evaluated

legend 	 whether a legend is required in the plots or not, the default is legent=FALSE

main 	 the main title as character. If NULL the default title (shown the intervals) is shown

main.gsub 	 if the main.gsub (with default "@") appears in the main title then it is substituted with the default title.

ylab 	 the y-variable label

xlab 	 the x-variable label

ylim 	 the range of the y-variable axis

overlap 	 how much overlapping in the xvar intervals. Default value is overlap=0 for non overlapping intervals

save 	 whether to save the sample percentages or not with default equal to TRUE. In this case the functions produce a matrix giving the sample percentages for each interval

plot 	 whether to plot the centles. This option is usefull if the sample statistics only are to be used

... 	 for extra arguments

Details

Value

Centile plots are produced and the sample centiles below each centile curve for each of the r ranges of x can be saved into a matrix.

Warning

This function is appropriate when only one continuous explanatory variable is fitted in the model

Note

Author(s)

Mikis Stasinopoulos, d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk, with contributions from Elaine Borghie

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss centiles, centiles.com

Examples


data(abdom)
h<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=BCT, data=abdom) 
mout <- centiles.split(h,xvar=abdom$x)
mout
rm(h,mout)

[Package gamlss version 1.7-9 Index]