coef.gamlss(gamlss)	R Documentation

Extract Model Coefficients in a GAMLSS fitted model

Description

coef.gamlss is the GAMLSS specific method for the generic function coef which extracts model coefficients from objects returned by modelling functions. `coefficients' is an alias for coef.

Usage


## S3 method for class 'gamlss':
coef(object, what = c("mu", "sigma", "nu", "tau"), ... )

Arguments

object 	 a GAMLSS fitted model

what 	 which parameter coefficient is required, default what="mu"

... 	 for extra arguments

Details

Value

Coefficients extracted from the GAMLSS model object.

Note

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk

References

RRigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, deviance.gamlss, fitted.gamlss

Examples


data(aids)
h<-gamlss(y~poly(x,3)+qrt, family=PO, data=aids) # 
coef(h)
rm(h)

[Package gamlss version 1.7-9 Index]