computer(gamlss.data)	R Documentation

The Computer Failure Data files for GAMLSS

Description

computing: The data relate to DEC-20 computers which operated at the Open University in the 1980. They give the number of computers that broke down in each of the 128 consecutive weeks of operation, starting in late 1983, see Hand et al. (1994) page 109 data set 141.

Usage



data(computer)


Format

Data frames each with the following variable.

failure

a numeric vector showing the number of times computers failed

Details

Data sets usefull for the GAMLSS booklet

References

Hand et al. (1994) A handbook of small data sets. Chapman and Hall, London.

Examples


data(computer)
with(computer, plot(table(failure)))

[Package gamlss.data version 0.0-2 Index]