cs(gamlss)	R Documentation

Specify a Smoothing Cubic Spline Fit in a GAMLSS Formula

Description

The functions cs() scs() and vc are using cubic smoothing splines. They take a vector and return it with several attributes. The vector is used in the construction of the model matrix. The functions do no smoothing, but assigns the attributes to the vector to aid gamlss in the smoothing. The function doing the smoothing are gamlss.cs(), gamlss.scs() and gamlss.vc(). The gamlss.cs() is a modified version of the R function smooth.spline() which is used by the backfitting function additive.fit. The function gamlss.scs() just uses smooth.spline(). It differs from the function cs() in that allows cross valdation of the smoothing parameters. The (experimental) function vc can be use to fit varying coefficient models, see Hastie and Tibshirani (1993).

Usage


cs(x, df = 3, spar = NULL, c.spar = NULL)
vc(r, x, df = 3, spar = NULL, c.spar = NULL)
scs(x, df = NULL, spar = NULL, control.spar = NULL, all.knots = TRUE, nknots = NULL, penalty =1.4)

Arguments

x 	 the univariate predictor, (or expression, that evaluates to a numeric vector). For the function vc the x argument is the vector which has its (linear) coefficient change with r

df 	 the desired equivalent number of degrees of freedom (trace of the smoother matrix minus two for the constant and linear fit). The real smoothing parameter (spar below) is found such that df=tr(S)-2, where S is the implicit smoother matrix. Values for df should be greater than 0, with 0 implying a linear fit.

spar 	 smoothing parameter, typically (but not necessarily) in (0,1]. The coefficient lambda of the integral of the squared second derivative in the fit (penalized log likelihood) criterion is a monotone function of `spar', see the details in smooth.spline.

c.spar 	 This is an option to be used when the degrees of freedom of the fitted gamlss object are different from the ones given as input in the option df. The default values used are the ones given the option control.spar in the R function smooth.spine() and they are c.spar=c(-1.5, 2). For very large data sets e.g. 10000 observations, the upper limit may have to increase for example to c.spar=c(-1.5, 2.5). Use this option if you have received the warning 'The output df are different from the input, change the control.spar'. c.spar can take both vectors or lists of length 2, for example c.spar=c(-1.5, 2.5) or c.spar=list(-1.5, 2.5) would have the same effect.

r 	 for the function vc, r represent the vector of the explanatory variable which effects the coefficients of x i.e. beta(r)*x. Both the x and r vectors should be adjusted by subtracting the their mean

control.spar 	 see above c.spar or the equivalent argument in the function smooth.spline.

all.knots 	 if TRUE, all distinct points in x are used as knots. If FALSE (default), a subset of x is used, see the same argument for function smooth.spline

nknots 	 integer giving the number of knots to use when all.knots=FALSE. Per default, this is less than n, the number of unique x values for n > 49

penalty 	 the penalty applied to cross validation

Details

Note that cs itself does no smoothing; it simply sets things up for the function gamlss() which in turn uses the function additive.fit() for backfitting which in turn uses gamlss.cs()

Note that cs() and s() functions behave differently at their default values that is if df and lambda are not specified. cs(x) by default will use 3 extra degrees of freedom for smoothing for x. ps(x) by default will estimate lambda (and the degrees of freedom) automatically using generalised cross validation (GCV). Note that if GCV is used the convergence of the gamlss model can be less stable compared to a model where the degrees of freedom are fixed. This can be true especially for small data sets.

Value

the vector x is returned, endowed with a number of attributes. The vector itself is used in the construction of the model matrix, while the attributes are needed for the backfitting algorithms additive.fit(). Since smoothing splines includes linear fits, the linear part will be efficiently computed with the other parametric linear parts of the model.

Warning

For a user who wishes to compare the gamlss() results with the equivalent gam() results in S-plus: make sure when using S-plus that the convergence criteria epsilon and bf.epsilon in control.gam() are decreased sufficiently to ensure proper convergence in S-plus. Also note that the degrees of freedom are defined on top of the linear term in gamlss, but on top of the constant term in S-plus, (so use an extra degrees of freedom in S-plus in order to obtain comparable results to those in galmss).

Change the upper limit of spar if you received the warning 'The output df are different from the input, change the control.spar'.

For large data sets do not use expressions, e.g. cs(x^0.5) inside the gamlss function command but evaluate the expression, e.g. nx=x^0.5, first and then use cs(nx).

Note

The degrees of freedom df are defined differently from that of the gam() function in S-plus. Here df are the additional degrees of freedom excluding the constant and the linear part of x. For example df=4 in gamlss() is equivalent to df=5 in gam() in S-plus

Author(s)

Mikis Stasinopoulos and Bob Rigby

References

Hastie, T. J. and Tibshirani, R. J. (1993), Varying coefficient models (with discussion),J. R. Statist. Soc. B., 55, 757-796.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.cs, lo

Examples


# cubic splines example
data(aids)
# fitting a smoothing cubic spline with 7 degrees of freedom
# plus the a quarterly  effect  
aids1<-gamlss(y~cs(x,df=7)+qrt,data=aids,family=PO) # 
aids2<-gamlss(y~scs(x,df=5)+qrt,data=aids,family=PO) # 
aids3<-gamlss(y~scs(x)+qrt,data=aids,family=PO) # using GCV 
with(aids, plot(x,y))
lines(aids$x,fitted(aids1), col="red")
lines(aids$x,fitted(aids3), col="green")
rm(aids1, aids2, aids3)
# varying-coefficient example
data(rent)
attach(rent)
# adjusting the variables
Flbar<-Fl-mean(Fl)
Abar<-A-mean(A)
# additive model
 m1<-gamlss(R~cs(Flbar, df=3)+cs(Abar))
# varying-coefficient model
 m2<-gamlss(R~cs(Flbar, df=3)+cs(Abar)+vc(r=Abar,x=Flbar))
AIC(m1,m2)
detach(rent)

[Package gamlss version 3.0-0 Index]