enzyme(gamlss.mx)	R Documentation

Data used in gamlss.mx

Description

enzyme : The data comprise independent measurement of enzyme activity in the blood of 245 individuals. The data were analysed by Bechker at al. (1993).

brains : the brain size, brain, and body weight, body, for 28 differnt animals.

Usage


data(enzyme)
data(brains)

Format

enzyme : data frame with 245 observations on the following variable act.

brains : data frame with 28 observations on the following variables. body, brain

act

a numeric vector showing enzyme activity in the blood of 245 individuals.

body

a numeric vector showing the body weight of 28 differnt animals

brain

a numeric vector showing the brain size of 28 differnt animals

Details

Source

References

Examples


data(enzyme)
hist(enzyme$act)
data(brains)
brains$lbrain<-log(brains$brain)
brains$lbody<-log(brains$body)
with(brains, plot(lbrain~lbody))

[Package gamlss.mx version 1.5-2 Index]