fabric(gamlss.util)	R Documentation

The Fabric Data

Description

The data are 32 observations on faults in rolls of fabric

Usage

data(fabric)

Format

A data frame with 32 observations on the following 3 variables.

leng

the length of the roll : a numeric vector

y

the number of faults in the roll of fabric : a discrete vector

x

the log of the length of the roll : a numeric vector

Details

The data are 32 observations on faults in rolls of fabric taken from Hinde (1982) who used the EM algorithm to fit a Poisson-normal model. The response variable is the number of faults in the roll of fabric and the explanatory variable is the log of the length of the roll.

Source

John Hinde

References

Hinde, J. (1982) Compound Poisson regression models: in GLIM 82, Proceedings of the International Conference on Generalized Linear Models, ed. Gilchrist, R., 109–121, Springer: New York.

Examples


data(fabric)
attach(fabric)
plot(x,y)
detach(fabric)

[Package gamlss.util version 0.0-2 Index]