find.hyper(gamlss)	R Documentation

A function to select values of hyperparameters in a GAMLSS model

Description

This function selects the values of hyper parameters and/or non-linear parameters in a GAMLSS model. It uses the R{} function optim which then minimised the generalized Akaike information criterion (GAIC) with a user defined penalty.

Usage


find.hyper(model = NULL, parameters = NULL, other = NULL, penalty = 2,
        steps = c(0.1), lower = -Inf, upper = Inf, method = "L-BFGS-B",
        ...)

Arguments

model 	 this is a GAMLSS model. e.g.
model=gamlss(y~cs(x,df=p[1]),sigma.fo=~cs(x,df=p[2]),data=abdom) where p[1] and p[2] denore the parametets to be estimated

parameters 	 the starting values in the search of the optimum hyperparameters and/or non-linear parameters e.g. parameters=c(3,3)

other 	 this is used to optimize other non-parameters, for example a transformation of the explanatory variable of the kind x^p[3], others=quote(nx<-x^p[3]) where nx is now in the model formula

penalty 	 specifies the penalty in the GAIC, (the default is 2 e.g. penalty=3

steps 	 the steps taken in the optimization procedure [see the ndeps option in optim()], by default is set to 0.1 for all hyper parameters and non-linear parameters

lower 	 the lower permissible level of the parameters i.e. lower=c(1,1) this does not apply if a method other than the default method "L-BFGS-B" is used

upper 	 the upper permissible level of the parameters i.e. upper=c(30,10), this is not apply if a method other than the default method "L-BFGS-B" is used

method 	 the method used in optim() to numerically minimize the GAIC over the hyperparameters and/or non-linear parameters. By default this is "L-BFGS-B" to allow box-restriction on the parameters

... 	 for extra arguments to be passed to the R function optim() used in the optimization

Details

This is an experimental function which appears to work well for the search of the optimum degrees of freedom and non-linear parameters (e.g. power parameter lambda used to transform x to x^lambda). Further investigation will check whether this function is reliable in general.

Value

The function turns the same output as the function optim()

par 	 the optimum hyperparameter values

value 	 the minimized value of the GAIC

counts 	 A two-element integer vector giving the number of calls to `fn' and `gr' respectively

convergence 	 An integer code. `0' indicates successful convergence. see the function optim() for other errors

message 	 A character string giving any additional information returned by the optimizer, or `NULL'

Warning

It may be slow to find the optimum

Author(s)

Mikis Stasinopoulos

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, plot.gamlss, optim

Examples


data(abdom)
attach(abdom)
# declare the model
mod1<-quote(gamlss(y~cs(nx,df=p[1]),family=BCT,data=abdom,
                        control=gamlss.control(trace=FALSE)))
# we want also to check for a transformation in x 
# so we use the other option
op<-find.hyper(model=mod1, other=quote(nx<-x^p[2]), parameters=c(3,0.5), 
            lower=c(1,0.001), steps=c(0.1,0.001))
# the optimum parameters found are 
# p=(p[1],p[2]) = (2.944836 0.001000) = (df for mu, lambda)
# so it needs df = 3 on top of the constant and linear 
# in  the cubic spline model for mu since p[1] is approximately  3
# and log transformation for x since p[2] is approximately  0 
op
rm(op)

[Package gamlss version 1.9-1 Index]