gamlss.control(gamlss)	R Documentation

Auxiliary for Controlling GAMLSS Fitting

Description

Auxiliary function as user interface for gamlss fitting. Typically only used when calling gamlss function with the option control.

Usage


gamlss.control(c.crit = 0.001, n.cyc = 20, mu.step = 1, sigma.step = 1, nu.step = 1, 
               tau.step = 1, gd.tol = 5, iter = 0, trace = TRUE, autostep = TRUE, 
               save = TRUE, ...)


Arguments

c.crit 	 the convergence criterion for the algorithm

n.cyc 	 the number of cycles of the algorithm

mu.step 	 the step length for the parameter mu

sigma.step 	 the step length for the parameter sigma

nu.step 	 the step length for the parameter nu

tau.step 	 the step length for the parameter tau

gd.tol 	 global deviance tolerance level

iter 	 starting value for the number of iterations, typically set to 0 unless the function refit is used

trace 	 whether to print at each iteration (TRUE) or not (FALSE)

autostep 	 whether the steps should be halved automatically if the new global deviance is greater that the old one, the default is autostep=TRUE

save 	 save=TRUE, (the default), saves all the information on exit. save=FALSE saves only limited information as the global deviance and AIC. For example fitted values, design matrices and additive terms are not saved. The latest is useful when gamlss() is called several times within a procedure.

... 	 for extra arguments

Details

The step length for each of the parameters mu, sigma, nu or tau is very useful to aid convergence if the parameter has a fully parametric model. However using a step length is not theoretically justified if the model for the parameter includes one or more smoothing terms, (even thought it may give a very approximate result).

The c.crit can be increased to speed up the convergence especially for a large set of data which takes longer to fit. When `trace' is TRUE, calls to the function cat produce the output for each outer iteration.

Value

A list with the arguments as components.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss

Examples


data(aids)
h<-gamlss(y~poly(x,3)+qrt, family=PO, data=aids) # 
con<-gamlss.control(mu.step=0.1)
h<-gamlss(y~poly(x,3)+qrt, family=PO, data=aids, control=con) # 
rm(h,con)

[Package gamlss version 1.9-9 Index]