gamlss.cs(gamlss)	R Documentation

Support for Function cs() s() and vc()

Description

This is support for the functions cs(), s() and vc(). It is not intended to be called directly by users. The function gamlss.cs is based on the R function smooth.spline

Usage


gamlss.cs(x, y = NULL, w = NULL, df = 5, spar = NULL, cv = FALSE,  nknots = NULL,
         all.knots = TRUE, df.offset = 0, penalty = 1, 
         control.spar = list(low = -1.5, high = 2), xeval = NULL)
gamlss.scs(x, y = NULL, w = NULL, df = NULL, spar = NULL, cv = FALSE, all.knots = FALSE, 
          nknots = NULL, keep.data = TRUE, df.offset = 0, penalty = 1.4, 
          control.spar = list(low = -1.5, high = 2), xeval = NULL)
gamlss.vc(x, y = NULL, w = NULL, df = 5, spar = NULL, cv = FALSE, all.knots = TRUE, 
          df.offset = 0, penalty = 1, control.spar = list(low = -1.5, high = 2), 
          xeval = NULL)

Arguments

x 	 the design matrix

y 	 the response variable

w 	 prior weights

df 	 effective degrees of freedom

spar 	 spar the smoothing parameter

cv 	 options for the smooth.spline function not to use here

nknots 	 options for the smooth.spline function not to use here

all.knots 	 options for the smooth.spline function not to use here

df.offset 	 options for the smooth.spline function not to use here

keep.data 	 whether to keep data

penalty 	 options for the smooth.spline function not to use here

control.spar 	 control for spar. It can be changed through cs

xeval 	 used in prediction

Value

Returns a class "smooth.spline" object with

residuals 	 The residuals of the fit

fitted.values 	 The smoothing values

var 	 the variance for the fitted smoother

lambda 	 the final value for spar

nl.df 	 the smoothing degrees of freedom excluding the constant and linear terms, i.e. (df-2)

coefSmo 	 this is a list containing among others the knots and the coefficients

...

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

See Also

gamlss, cs

[Package gamlss version 2.0-0 Index]