gamlss.fp(gamlss)	R Documentation

Support for Function fp()

Description

Those are support for the functions fp() and pp. It is not intended to be called directly by users.

Usage


gamlss.fp(x, y, w, npoly = 2, xeval = NULL)
gamlss.pp(x, y, w)

Arguments

x 	 the x for function gamlss.fp is referred to the design matric of the specific parameter model (not to be used by the user)

y 	 the y for function gamlss.fp is referred to the working variable of the specific parameter model (not to be used by the user)

w 	 the w for function gamlss.fp is referred to the iterative weight variable of the specific parameter model (not to be used by the user)

npoly 	 a positive indicating how many fractional polynomials should be considered in the fit. Can take the values 1, 2 or 3 with 2 as default

xeval 	 used in prediction

Value

Returns a list with

fitted.values 	 fitted

residuals 	 residuals

var

nl.df 	 the trace of the smoothing matrix

lambda 	 the value of the smoothing parameter

coefSmo 	 the coefficients from the smoothing fit

varcoeff 	 the variance of the coefficients

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, fp

[Package gamlss version 1.9-9 Index]