gamlss.lo(gamlss)	R Documentation

Support for Function lo()

Description

This is support for the loess function lo(). It is not intended to be called directly by users. The function gamlss.lo is based on the R function loess

Usage


gamlss.lo(x, y, w = NULL, span, df = NULL, degree = 1, ncols =
                 FALSE, wspan = TRUE, parametric = FALSE, drop.square
                 = FALSE, normalize = FALSE, family = "gaussian",
                 method = "loess", control = loess.control(...), xeval
                 = NULL, ...)

Arguments

x 	 the design matrix

y 	 the response variable

w 	 prior weights

span 	 the smoothing parameter

df 	 effective degrees of freedom

degree 	 the order of the polynomial

ncols 	 the number of columns of the x matrix

wspan 	 argument for the loess function not to use here

parametric 	 argument for the loess function not to use here

drop.square 	 argument for the loess function not to use here

normalize 	 argument for the loess function not to use here

family 	 argument for the loess function not to use here

method 	 argument for the loess function not to use here

control 	 argument for the loess function not to use here

xeval 	 used in prediction

... 	 further arguments passed to or from other methods.

Value

Returns an object with

fitted 	 the smooth values

residuals 	 the residuals

var 	 the variance of the smoother

nl.df 	 the non-linear degrees of freedom

coefSmo 	 with value NULL

lambda 	 the value of span

Author(s)

Mikis Stasinopoulos based on Brian Ripley loess function in R

See Also

gamlss, lo

[Package Contents]