gamlss-nl-package(gamlss.nl)	R Documentation

The GAMLSS add on package for fiting parametric non linear models

Description

The main purpose of this package is to allow non-linear fitting within a GAMLSS model. The main function nlgamlss() can fit any parametric (up to four distribution parameters) GAMLSS family of distributions.

Details

Package: 	 gamlss-nl

Type: 	 Package

Version: 	 1.5.0

Date: 	 2005-12-12

License: 	 GPL (version 2 or later)

Author(s)

Mikis Stasinopoulos <d.stasinopoulos@londonmet.ac.uk>, Bob Rigby <r.rigby@londonmet.ac.uk> based on work of Jim Lindsey and Philippe Lambert.

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape, (with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

See Also

gamlss

Examples


data(la)
# fitting the Johnson's Su distribtion to the data
modJSU   <- nlgamlss(y=PET60, mu.fo= ~bflow*(1-p1*exp(-p2/bflow)), sigma.formula=~1, 
                       nu.fo=~1, mu.start = c(.6, 110), sigma.start= 3, nu.start=1, 
                       tau.start=0.6, family=JSU, data=la) 
plot(modJSU)
summary(modJSU)
vcov(modJSU)

[Package gamlss.nl version 1.7-0 Index]