gamlss.ps(gamlss)	R Documentation

Support for Functions ps() pb() cy() and ri()

Description

Those functions are support for the functions ps() and pb(). The functions are not intended to be called directly by users.

Usage


gamlss.ps(x, y, w, xeval = NULL, ...)
gamlss.pb(x, y, w, xeval = NULL, ...)
gamlss.ridge(x, y, w, xeval = NULL, ...)
gamlss.ri(x, y, w, xeval = NULL, ...)
gamlss.cy(x, y, w, xeval = NULL, ...)

Arguments

x 	 the x for function gamlss.fp is referred to the design matric of the specific parameter model (not to be used by the user)

y 	 the y for function gamlss.fp is referred to the working variable of the specific parameter model (not to be used by the user)

w 	 the w for function gamlss.fp is referred to the iterative weight variable of the specific parameter model (not to be used by the user)

xeval 	 used in prediction

... 	 further arguments passed to or from other methods.

Value

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, ps, pb

[Package gamlss version 1.9-7 Index]