gamlss.ra(gamlss)	R Documentation

Support for Function ra()

Description

This is support for the random effect smoother function ra() . It is not intended to be called directly by users. The function gamlss.ra is similar to the GAMLSS function gamlss.random. Both functions can be used with the same effect.

Usage


gamlss.ra(x, y, w, df = NULL)

Arguments

x 	 the explanatory design matrix

y 	 the response variable

w 	 iterative weights

df 	 effective degrees of freedom

Details

Value

Returns an list with

fitted.values 	 fitted values

residuals 	 residuals

var 	 variances of the fitted values

nl.df 	 the trace of the smoothing matrix

lambda 	 the value of the smoothing parameter

coefSmo 	 the coefficients from the smoothing fit

varcoeff 	 the variance of the coefficients

Note

This is an experimental function and should be used with care

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

RRigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, ra, random

[Package gamlss version 1.8-2 Index]