gamlss.random(gamlss)	R Documentation

Support for Function random()

Description

This is support for the function random(). It is not intended to be called directly by users. The function gamlss.radom is similar to the GAMLSS function gamlss.ra.

Usage


gamlss.random(x, y, w, df = NULL, lambda = 0)

Arguments

x 	 the explanatory design matrix

y 	 the response variable

w 	 iterative weights

df 	 effective degrees of freedom

lambda 	 the smoothing parameter

Value

Returns a list with

y 	 the fitted values

residuals 	 the residuals

var 	 the variance of the fitted values

lambda 	 the final lambda, the smoothing parameter

coefSmo 	 with value NULL

Author(s)

Mikis Stasinopoulos, based on Trevor Hastie function gam.random

References

Chambers, J. M. and Hastie, T. J. (1991). Statistical Models in S, Chapman and Hall, London.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, random, ra

[Package gamlss version 1.8-2 Index]