gamlss.rc(gamlss)	R Documentation

Support for Function rc()

Description

This is support for the function rc(). It is not intended to be called directly by users.

Usage


gamlss.rc(x, y, w, df = NULL)

Arguments

x 	 the explanatory design matrix

y 	 the response variable

w 	 iterative weights

df 	 effective degrees of freedom

Value

Returns a list with

fitted.values 	 fitted

residuals 	 residuals

var

nl.df 	 the trace of the smoothing matrix

lambda 	 the value of the smoothing parameter

coefSmo 	 the coefficients from the smoothing fit

varcoeff 	 the variance of the coefficients

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, rc

[Package gamlss version 1.9-9 Index]