gamlss.tr-package(gamlss.tr)	R Documentation

The GAMLSS Add On Package for Truncated Distributions

Description

The main purpose of this package is to allow the user of the GAMLSS models to fit truncated distributions.

Details

Package: 	 gamlss.tr

Type: 	 Package

Version: 	 1.0

Date: 	 2005-11-21

License: 	 GPL (version 2 or later)

~~ An overview of how to use the package, including the most important functions ~~

Author(s)

Mikis Stasinopoulos <d.stasinopoulos@londonmet.ac.uk> and Bob Rigby <r.rigby@londonmet.ac.uk>

Maintainer: Mikis Stasinopoulos <d.stasinopoulos@londonmet.ac.uk>

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

See Also

gamlss,gamlss.family

Examples


gen.trun(par=c(0,100),family="TF", name="0to100", type="both")
op <- par(mfrow = c(2, 2)) 
plot(function(x) dTF0to100(x, mu=80 ,sigma=20, nu=5), 0, 100)
plot(function(x) pTF0to100(x, mu=80 ,sigma=20, nu=5), 0, 100)
plot(function(x) qTF0to100(x, mu=80 ,sigma=20, nu=5), 0.01, .999)
hist(rTF0to100(1000, mu=80 ,sigma=20, nu=5))
op

[Package gamlss.tr version 1.8-1 Index]