gen.cens(gamlss.cens)	R Documentation

A Function to Generate Appropriate Functions to Be Used to Fit a Censored Response variable in GAMLSS

Description

The gen.cens() function allows the user to generate a d, p, (dummy) q and fitting gamlss functions for censor and interval response variables. The function can take any gamlss.family distribution.

Usage


gen.cens(family = "NO", name = "cens", 
          type = c("right", "left", "interval"), ...)

Arguments

family 	 a gamlss.family object, which is used to define the distribution and the link functions of the various parameters. The distribution families supported by gamlss() can be found in gamlss.family and in the package gamlss.dist.

name 	 the characters you want to add to the name of new functions, by default is the first letter of type and c i.e WEIic for WEI (weibull) interval response variable

type 	 whether right, left or in interval censoring is required, (right is the default)

... 	 for extra arguments

Details

Value

Returns the d, p, (dummy) q and the fitting used in the fitting gamlss algorithm (The one used in the fitting gamlss algorithm) of a gamlss.family distribution.

Note

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk and Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

cens.d, cens.p, cens

Examples


library(gamlss.dist)
data(lip)
gen.cens(WEI,type="interval") 
WEIic
gg1<- gamlss(y ~ poly(Tem,2)+poly(pH,2)+poly(aw,2), data=lip, family=WEIic, c.crit=0.00001, n.cyc=200, trace=FALSE)

[Package gamlss.cens version 1.8.1 Index]