glass(gamlss.util)	R Documentation

The Glass Data files for GAMLSS

Description

glass: show the strength of glass fibres, measured at the National Physical Laboratory, England, see Smith and Naylor (1987), (the unit of measurement were not given in the paper).

Usage


data(glass)

Format

Data frames each with the following variable.

strength

a numeric vector showing the strength of glass fibres

Details

Data sets usefull for the GAMLSS booklet

References

Smith R. L. Naylor, J. C. (1987) A comparison of maximum likelihood and Bayesian estimators for the three-parameter Weibull distributuion. Appl. Statist. 36, 358-369

Examples


data(glass)
with(glass, hist(strength))

[Package gamlss.util version 0.0-2 Index]