glim.control(gamlss)	R Documentation

Auxiliary for Controlling the inner algorithm in a GAMLSS Fitting

Description

Auxiliary function used for the inner iteration of gamlss algorithm. Typically only used when calling gamlss function through the option i.control.

Usage


glim.control(cc = 0.001, cyc = 50,  glm.trace = FALSE, 
             bf.cyc = 30, bf.tol = 0.001, bf.trace = FALSE, 
             ...)

Arguments

cc 	 the convergence criterion for the algorithm

cyc 	 the number of cycles of the algorithm

glm.trace 	 whether to print at each iteration (TRUE) or not (FALSE)

bf.cyc 	 the number of cycles of the backfitting algorithm

bf.tol 	 the convergence criterion (tolerance level) for the backfitting algorithm

bf.trace 	 whether to print at each iteration (TRUE) or not (FALSE, the default)

... 	 for extra arguments

Value

A list with the arguments as components

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape, (with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss

Examples


data(aids)
con<-glim.control(glm.trace=TRUE)
h<-gamlss(y~poly(x,3)+qrt, family=PO, data=aids, i.control=con) # 
rm(h,con)


[Package gamlss version 1.7-9 Index]