hodges(gamlss.util)	R Documentation

Hodges data

Description

There two data sets contain data used in Hodges (1998). In addition to the data used in that manuscript, it contains other data items.

The original data consists of two matrices of dimensions of 341x6 and a 45x4 respectively.

The first matrix hodges describes plans. The information for each plan is: the state, a two-character code that identifies plans within state, the total premium for an individual, the total premium for a family, the total enrollment of federal employees as individuals, and the total enrollment of federal employees as families.

The second matrix, hodges, describes states. The information for each state is: its two-letter abbreviation, the state average expenses per admission (from American Medical Association 1991 Annual Survey of Hospitals), population (1990 Census), and the region (from the Marion Merrill Dow Managed Care Digest 1991).

The Hodges manuscript used these variables: Plan level: individual premium, individual enrollment. State level: expenses per admission, region.

Usage

data(hodges)

Format

Two data frames the first with 341 observations on the following 6 variables.

state

a factor with 45 levels AL AZ CA CO CT DC DE FL GA GU HI IA ID IL IN KS KY LA MA MD ME MI MN MO NC ND NE NH NJ NM NV NY OH OK OR PA PR RI SC TN TX UT VA WA WI

plan

a two-character code that identifies plans within state declared here as factor with 325 levals.

prind

a numeric vector showing the total premium for an individual

prfam

a numeric vector showing the total premium for a family

enind

a numeric vector showing the total enrollment of federal employees as individuals

enfam

a numeric vector showing the total enrollment of federal employees as families.

and the second with 45 observations on the following 4 variables

State

a factor with levels same as state above

expe

a numeric vector showing the state average expenses per admission (from American Medical Association 1991 Annual Survey of Hospitals)

pop

a numeric vector shoing the population (1990 Census)

region

the region (from the Marion Merrill Dow Managed Care Digest 1991), a factor with levels MA MT NC NE PA SA SC

Source

http://www.biostat.umn.edu/~hodges/

References

Hodges, J. S. (1998). Some algebra and geometry for hierarchical models, applied to diadnostics. J. R. Statist. Soc. B., 60 pp 497:536.

Examples


data(hodges)
attach(hodges)
plot(prind~state, cex=1, cex.lab=1.5, cex.axis=1, cex.main=1.2)
str(hodges)
data(hodges1)
str(hodges1)

[Package gamlss.util version 0.0-2 Index]