lice(gamlss.data)	R Documentation

Data files for GAMLSS

Description

lice : The data come from Williams (1944) (also used by Stein and Juritz (1988).) and they are lice per head of Hindu male prisoners in Cannamore, South India, 1937-1939.

Usage



data(lice)


Format

Data frames each with the following variable.

head

a numeric vector showing the number lice per head of Hindu male prisoners in Cannamore, South India, 1937-1939.

freq

a numeric vector showing the frequency of lice per head

Details

Data sets usefull for the GAMLSS booklet

References

Stein, G. Z. and Juritz, J. M. (1988). Linear models with an inverse Gaussian-Poisson error distribution. Communications in Statistics- Theory and Methods, 17, 557-571.

Examples


data(lice)

[Package gamlss.data version 0.0-2 Index]