lip(gamlss.cens)	R Documentation

Data for lip

Description

The data set used in this package are collected by Dr Peggy Braun (University of Leipzig) and passed on to use by professor Jane Sutherland of London Metropolitan University.

It consists of experimental enzymology results from a research project which attempted to develop a generic food spoilage model.

The data set contains a column called NAMES, which shows the experiment name, three columns with values of the environmental conditions: temperature (Tem), pH and water activity (aw), and the rest of the columns contains the activity of the cocktails, observed at certain days.

The researchers recorded the activity of proteases and lipases in each cocktail and were interested in predicting the time when the activity started given the environmental conditions. The activity is a positive integer and enzymes are considered inactive when activity=0.

Usage

data(lip)

Format

A data frame with 120 observations on the following 14 variables.

name

a factor with levels the different experiment

Tem

a numeric vector showing the temperature

pH

a numeric vector PH

aw

a numeric vector water activity

X0.d

a numeric vector if enzyme reacted at day 0

X1.d

a numeric vector if enzyme reacted at day 1

X2.d

a numeric vector if enzyme reacted at day 2

X4.d

a numeric vector if enzyme reacted at days 3 or 4

X11.d

a numeric vector if enzyme reacted at days 5 to 11

X18d.

a numeric vector if enzyme reacted at days 12 to q18

X25.d

a numeric vector if enzyme reacted at days 19 to 25

X32.d

a numeric vector if enzyme reacted at days 26 to 32

X39.d

a numeric vector if enzyme reacted at days 33 to 39

y

a matrix with 3 columns: this is a Surv() object indicating the start the finish and censored indicator as defined in function Surv() of survival.

Details

Source

Prof. Jane Sutherland, London Metropolitan University

References

Examples


data(lip)
with(lip, y)

[Package gamlss.cens version 1.8.1 Index]