margolin(gamlss.data)	R Documentation

The Margolin Data files for GAMLSS

Description

margolin: Margolin et al. (1981) present data from an Ames Salmonella assay, where y is the number of revertant colonies observed on a plate given a dose y of quinoline. The data were subsequently analysed by Breslow (1984), Lawless (1987) and Saha and Paul (2005).

Usage


data(margolin)

Format

Data frames each with the following variable.

y

a numeric vector showing the number of revertant colonies observed on a plate given a dose x of quinoline.

x

a numeric vector showing a a dose x of quinoline.

Details

Data sets usefull for the GAMLSS booklet

References

Breslow, N. (1984) Extra-Poisson variation in log-linear models. Applied Statistics, 33, 38-44.

Hand et al. (1994) A handbook of small data sets. Chapman and Hall, London.

Lawless, J.F. (1987) Negative binomial and mixed Poisson regression. The Canadian Journal of Statistics, 15, 209-225.

Margolin, B.H., Kaplan, N. and Zeiger, E. (1981) Statistical analysis of the Ames salmonella/microsome test. Proceedings of the National Academy of Science, U.S.A., 76, 3779-3783.

Saha, K. and Paul, S. (2005) Bias-Corrected Maximum Likelihood Estimator of the Negative Binomial Dispersion Parameter. Biometrics, 61, 179-185

Examples


data(margolin)
with(margolin, plot(y~x))

[Package gamlss.data version 0.0-2 Index]