model.frame.gamlss(gamlss)	R Documentation

Extract a model.frame a model matrix or terms from a GAMLSS object for a given distributional parameter

Description

model.frame.gamlss, model.matrix.gamlss and terms.gamlss are the gamlss versions of the generic functions model.frame, model.matrix and terms respectively.

Usage


## S3 method for class 'gamlss':
model.frame(formula, what = c("mu", "sigma", "nu", "tau"), ...)
## S3 method for class 'gamlss':
terms(x, what = c("mu", "sigma", "nu", "tau"), ...)
## S3 method for class 'gamlss':
model.matrix(object, what = c("mu", "sigma", "nu", "tau"), ...)

Arguments

formula 	 a gamlss object

x 	 a gamlss object

object 	 a gamlss object

what 	 for which parameter to extract the model.frame, terms or model.frame

... 	 for extra arguments

Details

Value

a model.frame, a model.matrix or terms

Note

Author(s)

Mikis Stasinopoulos

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss

Examples


data(aids)
mod<-gamlss(y~poly(x,3)+qrt, family=PO, data=aids) # 
model.frame(mod)
model.matrix(mod)
terms(mod, "mu")
rm(mod)

[Package gamlss version 1.7-9 Index]