par.plot(gamlss)	R Documentation

A function to plot parallel plot for repeated measurement data

Description

This function can be used to plot parallel plots for each individual in a repeated measurement study. It is based on the coplot() function of R.

Usage


par.plot(formula = NULL, data = NULL, subjects = NULL, 
          color = TRUE, show.given = TRUE, ...)

Arguments

formula 	 a formula describing the form of conditioning plot. A formula of the form y ~ x | a indicates that plots of y versus x should be produced conditional on the variable a. A formula of the form y ~ x| a * b indicates that plots of y versus x should be produced conditional on the two variables a and b.

data 	 a data frame containing values for any variables in the formula. By default the environment where par.plot was called from is used.

subjects 	 a factor which distinguish between the individual participants

color 	 whether the parallel plot are shown in colour, color=TRUE (the default) or not color=FALSE

show.given 	 logical (possibly of length 2 for 2 conditioning variables): should conditioning plots be shown for the corresponding conditioning variables (default 'TRUE')

... 	 for extra arguments

Details

Value

It returns a plot.

Note

Note that similar plot can be fount in the library nlme by Pinheiro and Bates

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss

Examples



library(nlme)
data(Orthodont)
par.plot(distance~age,data=Orthodont,sub=Subject)
par.plot(distance~age|Sex,data=Orthodont,sub=Subject)
par.plot(distance~age|Subject,data=Orthodont,sub=Subject,show.given=FALSE)


[Package gamlss version 1.7-9 Index]