plot.gamlss(gamlss)	R Documentation

Plot Residual Diagnostics for an GAMLSS Object

Description

This function provides four plots for checking the normalized (randomized for a discrete response distribution) quantile residuals of a fitted GAMLSS object, referred to as residuals below : a plot of residuals against fitted values, a plot of the residuals against an index or a specific explanatory variable, a density plot of the residuals and a normal Q-Q plot of the residuals. If argument ts=TRUE then the first two plots are replaced by the autocorrelation function (ACF) and partial autocorrelation function (PACF) of the residuals

Usage


## S3 method for class 'gamlss':
plot(x, xvar = NULL, parameters = NULL, ts = FALSE, 
           summaries = TRUE, ...)

Arguments

x 	 a GAMLSS fitted object

xvar 	 an explanatory variable to plot the residuals against

parameters 	 plotting parameters can be specified here

ts 	 set this to TRUE if ACF and PACF plots of the residuals are required

summaries 	 set this to FALSE if no summary statistics of the residuals are required

... 	 further arguments passed to or from other methods.

Details

This function provides four plots for checking the normalized (randomized) quantile residuals (called residuals) of a fitted GAMLSS object. Randomization is only performed for discrete response variables. The four plots are

residuals against the fitted values (or ACF of the residuals if ts=TRUE)
residuals against an index or specified x-variable (or PACF of the residuals if ts=TRUE)
kernel density estimate of the residuals
QQ-normal plot of the residuals

For time series response variables option ts=TRUE can be used to plot the ACF and PACF functions of the residuals.

Value

Returns four plots related to the residuals of the fitted GAMLSS model and prints summary statistics for the residuals if the summary=T

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Kalliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss

Examples



data(aids)
a<-gamlss(y~cs(x,df=7)+qrt,family=PO,data=aids)
plot(a)
rm(a)

[Package gamlss version 1.7-9 Index]