prof.term(gamlss)	R Documentation

Plotting the Profile: deviance or information criterion for one of the terms (or hyper-parameters) in a GAMLSS model

Description

This functions plots the profile deviance for a chosen parameter included in the linear predictor of any of the mu, sigma, nu or tau models so profile confidence intervals can be obtained. In can also be used to plot the profile of a specified information criterion for any hyperparameter.

Usage


prof.term(model = NULL, criterion = "GD", penalty = 2.5, other = NULL, 
         min = NULL, max = NULL, step = NULL, type = "o", xlabel = NULL, 
         plot = TRUE, term = TRUE, perc = 95, ... )

Arguments

model 	 this is a GAMLSS model, e.g.
model=gamlss(y~cs(x,df=this),sigma.fo=~cs(x,df=3),data=abdom), where this indicates the (hyper)parameter to be profiled

criterion 	 whether global deviance ("GD") or information criterion ("IC") is profiled. The default is criterion="GD"

penalty 	 The penalty value if information criterion is used in criterion, default penalty=2.5

other 	 this can be used to evaluate an expression before the actual fitting of the model

min 	 the minimum value for the parameter e.g. min=1

max 	 the maximum value for the parameter e.g. max=20

step 	 how often to evaluate the global deviance (defines the step length of the grid for the parameter) e.g. step=1

type 	 what type of plot required. This is the same as in type for plot, default value is type="o", that is, both line and points

xlabel 	 if a label for the axis is required

plot 	 whether to plot, plot=TRUE or save the results, plot=FALSE

term 	 this has the value TRUE and it should be changed to FALSE if a profile global deviance is required for a hyperparameter so the IC are suppressed

perc 	 what % confidence interval is required

... 	 for extra arguments

Details

This function can be use to provide likelihood based confidence intervals for a parameter involved in terms in the linear predictor(s). These confidence intervals are more accurate than the ones obtained from the parameters' standard errors. The function can also be used to plot a profile information criterion (with a given penalty) against a hyperparameter. This can be used to check the uniqueness in hyperparameter determination using for example find.df.

Value

Return a profile plot (if the argument plot=TRUE) or the values of the parameters and the IC or GD values otherwise

Warning

A dense grid (i.e. small step) evaluation of the global deviance can take a long time, so start with a sparse grid (i.e. large step) and decrease gradually the step length for more accuracy.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk and Bob Rigby r.rigby@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, prof.dev

Examples


data(aids)
gamlss(y~x+qrt,family=NBI,data=aids)
mod<-quote(gamlss(y ~ offset(this * x) + qrt, data = aids, family = NBI))
prof.term(mod, min=0.06, max=0.11, step=0.001)
mod1<-quote(gamlss(y ~ cs(x,df=this) + qrt, data = aids, family = NBI))
prof.term(mod1, min=1, max=15, step=1, criterion="IC")
mod2 <- quote(gamlss(y ~ x+I((x>this)*(x-this))+qrt,family=NBI,data=aids))
prof.term(mod2, min=1, max=45, step=1, criterion="GD")
rm(mod,mod1,mod2)

[Package gamlss version 1.7-9 Index]