ps(gamlss)	R Documentation

Specify a Penalised Beta (Regression) Spline Fit in a GAMLSS Formula

Description

There are several function operating penalised B-splines: pb(), cy(), and ps(). The functions take a vector and return it with several attributes. The vector is used in the construction of the design matrix X used in the fitting. The functions do not do the smoothing, but assign the attributes to the vector to aid gamlss in the smoothing. The functions doing the smoothing are gamlss.ps() gamlss.pb() and gamlss.cy() which are used in the backfitting function additive.fit. The function pb() is more efficient and faster than the original penalized smoothing function ps(). pb() allows the estimation of the smoothing parameters using different local (performance iterations) methods. The method are "ML", "ML-1", "EM", "GAIC" and "GCV". The function cy() fits a cycle penalised beta regression spline such as the last fitted value of the smoother is equal to the first fitted value.

Usage


pb(x, df = NULL, lambda = NULL, control = pb.control(...), ...)
pb.control(inter = 20, degree = 3, order = 2, start = 10, quantiles = FALSE, 
               method = c("ML", "GAIC", "GCV", "EM", "ML-1"), k = 2, ...)
cy(x, df = NULL, lambda = NULL, control = cy.control(...), ...)
cy.control(inter = 20, degree = 3, order = 2, start = 10, 
          method = c("ML", "GAIC", "GCV", "EM", "ML-1"), k = 2, ts=FALSE, ...)
ps(x, df = 3, lambda = NULL, ps.intervals = 20, degree = 3, order = 3)

Arguments

x 	 the univariate predictor

df 	 the desired equivalent number of degrees of freedom (trace of the smoother matrix minus two for the constant and linear fit)

lambda 	 the smoothing parameter

control 	 setting the control parameters

... 	 for extra arguments

inter 	 the no of break points (knots) in the x-axis

degree 	 the degree of the piecewise polynomial

order 	 the required difference in the vector of coefficients

start 	 the lambda starting value if the local methods are used, see below

quantiles 	 if TRUE the quantile values of x are use to determine the knots

ts 	 if TRUE assumes that it is a seasonal factor

method 	 The method used in the (local) performance iterations. Available methods are "ML", "ML-1", "EM", "GAIC" and "GCV"

k 	 the penalty used in "GAIC" and "GCV"

ps.intervals 	 the no of break points in the x-axis

Details

The ps() function is based on Brian Marx function which can be found in http://www.stat.lsu.edu/faculty/marx/. The pb() and cy() functions are based on Paul Eilers original R functions. Note that ps() and pb() functions behave differently at their default values if df and lambda are not specified. ps(x) by default will use 3 extra degrees of freedom for smoothing x. pb(x) by default will estimate lambda (and the degrees of freedom) automatically using a local method. Local (or performance iterations) methods available are (i) Maximum Likelihood, "ML", (ii) Generalized Akaike information criterion, "GAIC" (iii) Generalized Cross validation. (iv) EM-algorithm, "EM" (very slow) and (v) a modified version of the ML, "ML-1" which produce identical results with "EM" but faster. Note that if the local (or performance iterations) methods can be make the convergence of gamlss less stable compared to a model where the degrees of freedom are fixed.

Value

the vector x is returned, endowed with a number of attributes. The vector itself is used in the construction of the model matrix, while the attributes are needed for the backfitting algorithms additive.fit().

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Paul Eilers

References

http://www.stat.lsu.edu/faculty/marx/

Eilers, P. H. C. and Marx, B. D. (1996). Flexible smoothing with B-splines and penalties (with comments and rejoinder). Statist. Sci, 11, 89-121.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.ps, cs

Examples


data(aids)
# fitting a smoothing cubic spline with 7 degrees of freedom
# plus the a quarterly  effect  
aids1<-gamlss(y~ps(x,df=7)+qrt,data=aids,family=PO) # 
aids2<-gamlss(y~pb(x,df=7)+qrt,data=aids,family=PO) # 
aids3<-gamlss(y~pb(x)+qrt,data=aids,family=PO) #
with(aids, plot(x,y))
with(aids, lines(x,fitted(aids1),col="red"))
with(aids, lines(x,fitted(aids2),col="green"))
with(aids, lines(x,fitted(aids1),col="yellow"))
rm(aids1, aids2, aids3)
# simulate data
set.seed(555)
x = seq(0, 1, length = 100)
y = sign(cos(1 * x * 2 * pi + pi / 4)) + rnorm(length(x)) * 0.2
plot(y~x)
m1<-gamlss(y~cy(x))
lines(fitted(m1)~x)
rm(y,x,m1)

[Package gamlss version 3.0-1 Index]