rent(gamlss.util)	R Documentation

Rent data

Description

A survey was conducted in April 1993 by Infratest Sozialforschung. A random sample of accommodation with new tenancy agreements or increases of rents within the last four years in Munich was selected including: i) single rooms, ii) small apartments, iii) flats, iv) two-family houses. Accommodation subject to price control rents, one family houses and special houses, such as penthouses, were excluded because they are rather different from the rest and are considered a separate market. For the purpose of this study, 1967 observations of the variables listed below were used, i.e. the rent response variable R followed by the explanatory variables found to be appropriate for a regression analysis approach by Fahrmeir et al. (1994, 1995):

Usage

data(rent)

Format

A data frame with 1969 observations on the following 9 variables.

R

: rent response variable, the monthly net rent in DM, i.e. the monthly rent minus calculated or estimated utility cost

Fl

: floor space in square meters

A

: year of construction

Sp

: a variable indicating whether the location is above average, 1, (550 observations) or not, 0, (1419 observations)

Sm

: a variable indicating whether the location is below, 1, average (172 obs.) or not, 0, (1797 obs.)

B

: a factor with levels indicating whether there is a bathroom, 1, (1925 obs.) or not, 0, (44 obs.)

H

: a factor with levels indicating whether there is central heating, 1, (1580 obs.) or not, 0, (389 obs.)

L

: a factor with levels indicating whether the kitchen equipment is above average, 1, (161 obs.) or not, 0, (1808 obs.)

loc

: a factor (combination of Sp and Sm) indicating whether the location is below, 1, average, 2, or above average 3

Details

This set of data were used by Stasinopoulos et al. (2000) to fit a model where both the mean and the dispersion parameter of a Gamma distribution were modelled using the explanatory variables.

Source

Provide by Prof. L. Fahrmeir

References

Fahrmeir L., Gieger C., Mathes H. and Schneeweiss H. (1994) Gutachten zur Erstellung des Mietspiegels fur Munchen 1994, Teil B: Statistiche Analyse der Nettomieten. Hrsg: Landeshaupttstadt Munchen, Sozialreferat-Amt fur Wohnungswesen.

Fahrmeir L., Gieger C., and Klinger, A. (1995) Additive, dynamic and multiplicative regression. In Applied Statistics: Recent Developments, Vandenhoeck and Ruprecht, Gottingen.

Stasinopoulos, D. M., Rigby, R. A. and Fahrmeir, L., (2000), Modelling rental guide data using mean and dispersion additive models, Statistician, 49 , 479-493.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Examples


data(rent)
attach(rent)
plot(Fl,R)

[Package gamlss.util version 0.0-2 Index]