rqres.plot(gamlss)	R Documentation

Plotting Randomized Quantile Residuals

Description

This function plots QQ-plots of the normalized randomized quantile residuals (see Dunn and Smyth, 1996) for a model using a discrete GAMLSS family distribution.

Usage


rqres.plot(obj = NULL, howmany = 6, all = TRUE, save = FALSE, ...)

Arguments

obj 	 a fitted GAMLSS model object from a "discrete" type of family

howmany 	 The number of QQ-plots required up to ten i.e. howmany=6

all 	 if TRUE QQ-plots from howmany realizations are plotted. If FALSE then a single qq-plot of the median of the howmany realizations is plotted

save 	 If TRUE the median residuals can be saved

... 	 for extra arguments

Details

For discrete family distributions, the gamlss() function saves on exit one realization of randomized quantile residuals which can be plotted using the generic function plot which calls the plot.gamlss. Looking at only one realization can be misleading, so the current function creates QQ-plots for several realizations. The function allows up to 10 QQ-plots to be plotted. Occasionally one wishes to create a lot of realizations and then take a median of them (separately for each ordered value) to create a single median realization. The option all in combinations with the option howmany creates a QQ-plot of the medians of the normalized randomized quantile residuals. These 'median' randomized quantile residuals can be saved using the option (save=TRUE).

Value

If save it is TRUE then the vector of the median residuals is saved.

Warning

....

Note

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk

References

Dunn, P. K. and Smyth, G. K. (1996) Randomised quantile residuals, J. Comput. Graph. Statist., 5, 236–244

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

plot.gamlss, gamlss

Examples


data(aids) # fitting a model from a discrete distribution 
h<-gamlss(y~cs(x,df=7)+qrt, family=NBI, data=aids) # 
plot(h)
# plot qq- plots from 6 realization of the randomized quantile residuals
rqres.plot(h) 
# a qq-plot from the medians from 40 realizations
rqres.plot(h,howmany=40,all=FALSE) # 

[Package gamlss version 1.7-9 Index]