species(gamlss.util)	R Documentation

The Fish Species Data files for GAMLSS

Description

species: The number of different fish species (y=fish) was recorded for 70 lakes of the world together with explanatory variable x=log(lake) area. The data are given and analyzed by Stein and Juritz (1988).

Usage



data(species)


Format

Data frames each with the following variable.

fish

a numeric vector showing the number of different species in 70 lakes in the word

lake

a numeric vector showing the lake area

Details

Data sets usefull for the GAMLSS booklet

References

Stein, G. Z. and Juritz, J. M. (1988). Linear models with an inverse Gaussian-Poisson error distribution. Communications in Statistics- Theory and Methods, 17, 557-571.

Examples



data(species)
with(species, plot(fish~log(lake)))


[Package gamlss.util version 0.0-2 Index]