stylo(gamlss.util)	R Documentation

The Stylometric Data files for GAMLSS

Description

stylo : the data were given by Dr Mario Corina-Borja, see Chappas and Corina-Borja (2006), and has the number of a word appearing in a text.

Usage


data(stylo)

Format

Data frames each with the following variable.

word

a numeric vector showing the number a word appearing in a text

freq

a numeric vector showing the frequency of the number a word appearing in a text

Details

Data sets usefull for the GAMLSS booklet

References

Chappas C. and Corina-Borja M. A Stylometric analysis of newspapers periodical and news scriprs, Journal of Quantitative Linguistics, 13, 285-312

Examples


data(stylo)
plot(freq~word, type="h", data=stylo)

[Package gamlss.util version 0.0-2 Index]