tensile(gamlss.util)	R Documentation

The Tensile Data files for GAMLSS

Description

tensile: These data come from Quesenberry and Hales (1980) and were also reproduced in Hand et al. (1994), data set 180, page 140. They contain measurements of tensile strength of polyester fibres and the authors were trying to check if they were consistent with the lognormal distribution. According to Hand et al. (1994) "these data follow from a preliminary transformation. If the lognormal hypothesis is correct, these data should have been uniformly distributed".

Usage



data(tensile)


Format

Data frames each with the following variable.

str

a numeric vector showing the tensile strength

Details

Data sets usefull for the GAMLSS booklet

References

Hand et al. (1994) A handbook of small data sets. Chapman and Hall, London.

Quesenberry, C. and Hales, C. (1980). Concentration bands for uniformily plots. Journal of Statistical Computation and Simulation, 11, 41:53.

Examples



data(tensile)
with(tensile,hist(str))


[Package gamlss.util version 0.0-2 Index]