wp(gamlss)	R Documentation

Worm plot

Description

Provides single or multiple worm plots for GAMLSS fitted objects. This is a diagnostic tool for checking the residuals within different ranges (by default not overlapping) of the explanatory variable

Usage


wp(object, xvar = NULL, n.inter = 4, xcut.points = NULL, 
   overlap = 0, xlim.all = 4, xlim.worm = 3.5, 
   show.given = TRUE, line = TRUE, 
   ylim.all = 12 * sqrt(1/length(fitted(object))), 
   ylim.worm = 12 * sqrt(n.inter/length(fitted(object))), 
   cex = 1, pch = 21, ...)

Arguments

object 	 a GAMLSS fitted object

xvar 	 the explanatory variable against which the worm plots will be plotted

n.inter 	 the number of intervals in which the explanatory variable xvar will be cut

xcut.points 	 the x-axis cut off points e.g. c(20,30). If xcut.points=NULL then the n.inter argument is activated

overlap 	 how much overlapping in the xvar intervals. Default value is overlap=0 for non overlapping intervals

xlim.all 	 for the single plot, this value is the x-variable limit, default is xlim.all=4

xlim.worm 	 for multiple plots, this value is the x-variable limit, default is xlim.worm=3.5

show.given 	 whether to show the x-variable intervals in the top of the graph, default is show.given=TRUE

line 	 whether to plot the polynomial line in the worm plot, default value is line=TRUE

ylim.all 	 for the single plot, this value is the y-variable limit, default value is ylim.all=12*sqrt(1/length(fitted(object)))

ylim.worm 	 for multiple plots, this values is the y-variable limit, default value is ylim.worm=12*sqrt(n.inter/length(fitted(object)))

cex 	 the cex plotting parameter with default cex=1

pch 	 the pch plotting parameter with default pch=21

... 	 for extra arguments

Details

If the xvar argument is not specified then a single worm plot is used. In this case a worm plot is a detrended normal QQ-plot so departure from normality is highlighted. If the xvar is specified then we have as many worm plot as n.iter. In this case the x-variable is cut into n.iter intervals with an equal number observations and detrended normal QQ (i.e. worm) plots for each interval are plotted. This is a way of highlighting failures of the model within different ranges of the explanatory variable. The fitted coefficients from fitting cubic polynomials to the residuals (within each x-variable interval) can be obtain by e.g. coeffs<-wp(model1,xvar=x,n.iner=9). van Buuren et al. (2001) used these residuals to identify regions (intervals) of the explanatory variable within which the model does not fit adequately the data (called "model violation")

Value

For multiple plots the xvar intervals and the coefficients of the fitted cubic polynomials to the residuals (within each xvar interval) are returned.

Note

Author(s)

Mikis Stasinopoulos and Bob Rigby

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, 1-38.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

van Buuren and Fredriks M. (2001) Worm plot: simple diagnostic device for modelling growth reference curves. Statistics in Medicine, 20, 1259–1277

See Also

gamlss, plot.gamlss

Examples


data(abdom)
a<-gamlss(y~cs(x,df=3),sigma.fo=~cs(x,1),family=LO,data=abdom)
wp(a)
coeff1<-wp(a,abdom$x)
coeff1
rm(a,a1)

[Package gamlss version 1.7-9 Index]